change cnn_size_t to uint32_t and fix portability issue (#377)
* use cnn_size_t to fix portability issues

* declare cnn_size_t as uint32_t

* Fix errors

* Fix errors

+++ tiny_dnn/config.h
@@ -26,6 +26,7 @@
 */
 #pragma once
 #include <cstddef>
+#include <cstdint>
 
 /**
  * define if you want to use intel TBB library
@@ -95,6 +96,6 @@ typedef float float_t;
  * size of layer, model, data etc.
  * change to smaller type if memory footprint is severe
  **/
-typedef std::size_t cnn_size_t;
+typedef std::uint32_t cnn_size_t;
 
 }
